// src/types/FirstFiveFactors.d.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import { ContractOptions } from "web3-eth-contract";
import { EventLog } from "web3-core";
import { EventEmitter } from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface FirstFiveFactors extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): FirstFiveFactors;
  clone(): FirstFiveFactors;
  methods: {
    factorize(nums: number | string | BN): PayableTransactionObject<void>;

    firstFactor(): NonPayableTransactionObject<string>;

    secondFactor(): NonPayableTransactionObject<string>;

    thirdFactor(): NonPayableTransactionObject<string>;

    fourthFactor(): NonPayableTransactionObject<string>;

    fifthFactor(): NonPayableTransactionObject<string>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}